Add basic additional Python-only Layouts
