+++ nestedtensor/version.py
@@ -1,5 +1,3 @@
-__version__ = '0.0.1+de193f7'
-git_version = 'de193f7ba770000291e0e00bf0affdd7fd1d4353'
-from nestedtensor import _C
-if hasattr(_C, 'CUDA_VERSION'):
-    cuda = _C.CUDA_VERSION
+__version__ = '0.0.1+807ad4d'
+git_version = '807ad4d72332355fd1670003b180865e8d0fcc0d'
+USE_C_EXTENSION=False
